test: fix two test failures in observation parsing and journal-crud
Fix 1: Update LOINC 51855-5 test to pass `valueString` as dynamic
title instead of expecting a static 'Patient Note'. The source code
correctly uses `resource.valueString` as the patient's own subject
line — the test was wrong to assert a literal string.

Fix 2: Move `useDeleteJournal` test to its own file to avoid mock
interference. The top-level `vi.mock('@/services/api/record')` in
journal-crud.test.tsx was overriding the hook with `vi.fn()`,
making `result.current` undefined.
